feat: implement fadein component in landingpage
